Move anticheat, move example to different file
